// src/types/product.ts
export type ProductClassification = 'dietary_supplement' | 'medicine' | 'combo'
// ...
export interface Product {
  company_id: number
  product_id: number
  combo_id: number //combo
  category_id: Category
  product_code: string
  combo_code: string //combo
  product_name: string
  combo_name: string //combo
  representation_image: ImageId
  attachment_cloud_id: URLRes //combo
  image_ids: ImageId[]
  price_unit: number
  origin_price_unit: number
  sold_quantity: number
  uom_id: ProductUom
  create_date: Date
  attribute_minor_ids: AttributeMinor[]
  rel_uom_ids: ProductUom[]
  packaging_specifications: string
  promotion_category: string
  attribute_ids: ProductAttribute[]
  rel_attribute_ids: ProductRelAttribute[]
  has_variant: boolean
  star_rating: StarRatingRange
  rating_count: number
  product_available: number
  quantity: number
  price_list: PriceList
  stock_quantity: ProductUom
  product_type: ProductClassification
}

export interface PriceList {
  [productId: string]: {
    [uomId: string]: number
  }
}

export interface ProductUom {
  uom_id: number
  uom_name: string
  uom_full_standard_name?: string
  factor?: number
  quantity?: number
}

export interface AttributeId {
  attribute_id: number
  attribute_name: string
}

export interface ValueId {
  value_id: number
  value_name: string
}

export interface ProductAttribute {
  attribute: AttributeId
  attribute_value: ValueId
}

export interface ProductRelAttribute {
  attribute_id: AttributeId
  attribute_value: ValueId[]
}